+++ packages/components/src/ScalePanel/index.ts
@@ -1,6 +1,5 @@
 import ScalePanel from "./index.vue"
 export { ScalePanel }
 export * from "./types"
-export * from "./useHandle"
 export * from "./useScale"
 export * from "./types"