+++ packages/components/src/ThreePanel/index.ts
@@ -1,6 +1,5 @@
 import useDebugger from "./debugger/useDebugger"
-import useDialog from "./debugger/useDialog"
 import ThreePanel from "./index.vue"
 export type { ControlConf, CameraConf, RendererConf, ThreePanelConf } from "./index.vue"
 export * from "./three/index"
-export { useDebugger, useDialog, ThreePanel }
+export { useDebugger, ThreePanel }