+++ playground/src/routes.ts
@@ -1,9 +1,9 @@
+import { fetchVue, RouterTree } from "@toolmain/libs"
 import { createWebHistory, createRouter, type RouteRecordRaw } from "vue-router"
-import { localRoute } from "@toolmain/libs"
 
-export const initNode = new localRoute.tpl.RouterTree({
+export const initNode = new RouterTree({
   index: "/",
-}).resolve(localRoute.fetch.fetchVue())
+}).resolve(fetchVue())
 console.log(initNode.root)
 export const initRoutes = initNode.iter<RouteRecordRaw>(router => {
   return router as RouteRecordRaw